rename file, fix icon not centered, use specific gray color code, add some margin-top to loadmore button
